+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/annotation/AnnotationTextGenerator.java
@@ -19,7 +19,7 @@ final public class AnnotationTextGenerator {
 	}
 
 	/** @return the decorated {@link SmallVariantAnnotation} */
-	public SmallVariantAnnotation getAnnotation() {
+	public VariantAnnotation getAnnotation() {
 		return annotation;
 	}