+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/reference/SmallGenomeVariant.java
@@ -4,7 +4,7 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.ComparisonChain;
 
 import de.charite.compbio.jannovar.Immutable;
-import de.charite.compbio.jannovar.annotation.SmallVariantAnnotation;
+import de.charite.compbio.jannovar.annotation.VariantAnnotation;
 import de.charite.compbio.jannovar.impl.util.DNAUtils;
 
 // TODO(holtgrewe): We only want genome changes on the forward strand, make sure this does not lead to problems downstream.
@@ -178,7 +178,7 @@ public final class SmallGenomeVariant extends GenomeVariant {
 	}
 
 	@Override
-	public int compareTo(SmallVariantAnnotation other) {
+	public int compareTo(VariantAnnotation other) {
 		return ComparisonChain.start().compare(pos, other.getPos()).compare(ref, other.getRef())
 				.compare(alt, other.getAlt()).result();
 	}